[execution] Separate invariant check inputs from execution state
