+++ sui-execution/latest/sui-adapter/src/temporary_store.rs
@@ -13,6 +13,7 @@ use sui_protocol_config::ProtocolConfig;
 use sui_types::accumulator_event::AccumulatorEvent;
 use sui_types::accumulator_root::AccumulatorObjId;
 use sui_types::base_types::VersionDigest;
+use sui_types::coin_reservation::ParsedDigest;
 use sui_types::committee::EpochId;
 use sui_types::deny_list_v2::check_coin_deny_list_v2_during_execution;
 use sui_types::effects::{
@@ -27,7 +28,7 @@ use sui_types::inner_temporary_store::InnerTemporaryStore;
 use sui_types::object::Data;
 use sui_types::storage::{BackingStore, DenyListResult, PackageObject};
 use sui_types::sui_system_state::{AdvanceEpochParams, get_sui_system_state_wrapper};
-use sui_types::transaction::{GasData, TransactionKind};
+use sui_types::transaction::{Command, GasData, TransactionKind, is_gasless_transaction};
 use sui_types::{
     SUI_DENY_LIST_OBJECT_ID,
     base_types::{ObjectID, ObjectRef, SequenceNumber, SuiAddress, TransactionDigest},
@@ -44,6 +45,46 @@ use sui_types::{SUI_SYSTEM_STATE_OBJECT_ID, TypeTag, is_system_package};
 pub(crate) mod invariants;
 use invariants::InvariantChecker;
 
+#[derive(Default)]
+struct PostExecutionCheckInputs {
+    /// Per-`(address, type)` funds-accumulator reservation budget authorized by this transaction.
+    /// Shared by gasless execution validation and the post-execution invariant checks.
+    input_reservations: BTreeMap<(SuiAddress, TypeTag), u64>,
+    /// For the advance-epoch transaction, `(epoch_fees minted, epoch_rebates burned)`; `None`
+    /// for every other transaction. Needed by the expensive SUI conservation check.
+    advance_epoch_gas_summary: Option<(u64, u64)>,
+    /// The genesis transaction mints the initial SUI supply and so is exempt from conservation.
+    is_genesis: bool,
+    /// What each `Publish`/`Upgrade` command in the PTB says the package it writes should look like.
+    /// `None` when the transaction is not a PTB.
+    declared_packages: Option<Vec<(usize, BTreeSet<ObjectID>)>>,
+}
+
+impl PostExecutionCheckInputs {
+    fn new(
+        transaction: Option<(&TransactionKind, &GasData, SuiAddress)>,
+        enable_gasless: bool,
+    ) -> Self {
+        let Some((transaction_kind, gas_data, transaction_signer)) = transaction else {
+            return Self {
+                is_genesis: true,
+                ..Default::default()
+            };
+        };
+        Self {
+            input_reservations: compute_input_reservations(
+                transaction_kind,
+                gas_data,
+                transaction_signer,
+                enable_gasless,
+            ),
+            advance_epoch_gas_summary: transaction_kind.get_advance_epoch_tx_gas_summary(),
+            is_genesis: matches!(transaction_kind, TransactionKind::Genesis(_)),
+            declared_packages: declared_packages(transaction_kind),
+        }
+    }
+}
+
 pub struct TemporaryStore<'backing> {
     // The backing store for retrieving Move packages onchain.
     // When executing a Move call, the dependent packages are not going to be
@@ -53,6 +94,9 @@ pub struct TemporaryStore<'backing> {
     store: &'backing dyn BackingStore,
     tx_digest: TransactionDigest,
     input_objects: BTreeMap<ObjectID, Object>,
+    /// Immutable transaction-derived inputs needed for various checks after execution finishes.
+    // TODO: We should merge all input-derived immutable data to a single struct.
+    post_execution_check_inputs: PostExecutionCheckInputs,
 
     /// Store the original versions of the non-exclusive write inputs, in order to detect
     /// mutations (which are illegal, but not prevented by the type system).
@@ -67,8 +111,6 @@ pub struct TemporaryStore<'backing> {
     execution_results: ExecutionResultsV2,
     /// Objects that were loaded during execution (dynamic fields + received objects).
     loaded_runtime_objects: BTreeMap<ObjectID, DynamicallyLoadedObjectMetadata>,
-    /// A map from wrapped object to its container. Used during expensive invariant checks.
-    wrapped_object_containers: BTreeMap<ObjectID, ObjectID>,
     protocol_config: &'backing ProtocolConfig,
 
     /// Every package that was loaded from DB store during execution.
@@ -79,11 +121,6 @@ pub struct TemporaryStore<'backing> {
     /// any of the objects referenced in this set.
     receiving_objects: Vec<ObjectRef>,
 
-    /// The set of all generated object IDs from the object runtime during the transaction. This includes any
-    /// created-and-then-deleted objects in addition to any `new_ids` which contains only the set
-    /// of created (but not deleted) IDs in the transaction.
-    generated_runtime_ids: BTreeSet<ObjectID>,
-
     // TODO: Now that we track epoch here, there are a few places we don't need to pass it around.
     /// The current epoch.
     cur_epoch: EpochId,
@@ -92,16 +129,15 @@ pub struct TemporaryStore<'backing> {
     /// input objects. This allows us to commit them to the effects.
     loaded_per_epoch_config_objects: RwLock<BTreeSet<ObjectID>>,
 
-    /// Transaction-derived inputs and bookkeeping for the post-execution system-invariant checks
-    /// (SUI conservation, balance-accumulator authorization, object ownership). See
-    /// [`invariants::InvariantChecker`].
+    /// Execution-attempt bookkeeping for post-execution system checks.
     invariants: InvariantChecker,
 }
 
 impl<'backing> TemporaryStore<'backing> {
     /// Creates a new store associated with an authority store, and populates it with
     /// initial objects.
-    pub fn new(
+    #[allow(clippy::too_many_arguments)]
+    pub(crate) fn new(
         store: &'backing dyn BackingStore,
         input_objects: InputObjects,
         receiving_objects: Vec<ObjectRef>,
@@ -109,7 +145,10 @@ impl<'backing> TemporaryStore<'backing> {
         protocol_config: &'backing ProtocolConfig,
         cur_epoch: EpochId,
         _system_object_versions: BTreeMap<ObjectID, SequenceNumber>,
+        transaction: Option<(&TransactionKind, &GasData, SuiAddress)>,
     ) -> Self {
+        let post_execution_check_inputs =
+            PostExecutionCheckInputs::new(transaction, protocol_config.enable_gasless());
         let mutable_input_refs = input_objects.exclusive_mutable_inputs();
         let non_exclusive_input_original_versions = input_objects.non_exclusive_input_objects();
 
@@ -144,13 +183,12 @@ impl<'backing> TemporaryStore<'backing> {
             execution_results: ExecutionResultsV2::default(),
             protocol_config,
             loaded_runtime_objects: BTreeMap::new(),
-            wrapped_object_containers: BTreeMap::new(),
             runtime_packages_loaded_from_db: RwLock::new(BTreeMap::new()),
             receiving_objects,
-            generated_runtime_ids: BTreeSet::new(),
             cur_epoch,
             loaded_per_epoch_config_objects: RwLock::new(BTreeSet::new()),
-            invariants: InvariantChecker::new(),
+            post_execution_check_inputs,
+            invariants: InvariantChecker::default(),
         }
     }
 
@@ -572,12 +610,12 @@ impl<'backing> TemporaryStore<'backing> {
 
     pub fn drop_writes(&mut self) {
         self.execution_results.drop_writes();
-        // The PTB-emitted ranges pointed into the now-cleared accumulator_events vec.
-        self.invariants.clear();
+        self.invariants = InvariantChecker::default();
     }
 
     /// Consume this (post-execution) store and return the store used by a `BumpOnly` exit: keep
-    /// the input-derived state, discard everything execution produced, then bump the mutable
+    /// the input-derived state as well as information about the execution needed for replay,
+    /// discard everything related to the execution results, then bump the mutable
     /// inputs. Its effects record only those version bumps and the input dependencies.
     pub(crate) fn into_bump_only(self) -> Self {
         let Self {
@@ -592,13 +630,13 @@ impl<'backing> TemporaryStore<'backing> {
             receiving_objects,
             cur_epoch,
             protocol_config,
-            // Execution-derived - discarded.
+            post_execution_check_inputs,
+            // Represents what happened during execution, which needs to be kept.
+            loaded_runtime_objects,
+            runtime_packages_loaded_from_db,
+            loaded_per_epoch_config_objects,
+            // Execution outcomes, can be discarded.
             execution_results: _,
-            loaded_runtime_objects: _,
-            wrapped_object_containers: _,
-            runtime_packages_loaded_from_db: _,
-            generated_runtime_ids: _,
-            loaded_per_epoch_config_objects: _,
             invariants: _,
         } = self;
         let mut bump_only = Self {
@@ -612,13 +650,12 @@ impl<'backing> TemporaryStore<'backing> {
             receiving_objects,
             cur_epoch,
             protocol_config,
+            loaded_runtime_objects,
+            runtime_packages_loaded_from_db,
+            loaded_per_epoch_config_objects,
+            post_execution_check_inputs,
             execution_results: ExecutionResultsV2::default(),
-            loaded_runtime_objects: BTreeMap::new(),
-            wrapped_object_containers: BTreeMap::new(),
-            runtime_packages_loaded_from_db: RwLock::new(BTreeMap::new()),
-            generated_runtime_ids: BTreeSet::new(),
-            loaded_per_epoch_config_objects: RwLock::new(BTreeSet::new()),
-            invariants: InvariantChecker::new(),
+            invariants: InvariantChecker::default(),
         };
         // The only writes a BumpOnly exit records: bump the versions of the mutable inputs it locked.
         bump_only.ensure_active_inputs_mutated();
@@ -660,36 +697,12 @@ impl<'backing> TemporaryStore<'backing> {
         &mut self,
         wrapped_object_containers: BTreeMap<ObjectID, ObjectID>,
     ) {
-        #[cfg(debug_assertions)]
-        {
-            for (id, container1) in &wrapped_object_containers {
-                if let Some(container2) = self.wrapped_object_containers.get(id) {
-                    assert_eq!(container1, container2);
-                }
-            }
-            for (id, container1) in &self.wrapped_object_containers {
-                if let Some(container2) = wrapped_object_containers.get(id) {
-                    assert_eq!(container1, container2);
-                }
-            }
-        }
-        // Merge the two maps because we may be calling the execution engine more than once
-        // (e.g. in advance epoch transaction, where we may be publishing a new system package).
-        self.wrapped_object_containers
-            .extend(wrapped_object_containers);
+        self.invariants
+            .save_wrapped_object_containers(wrapped_object_containers);
     }
 
     pub fn save_generated_object_ids(&mut self, generated_ids: BTreeSet<ObjectID>) {
-        #[cfg(debug_assertions)]
-        {
-            for id in &self.generated_runtime_ids {
-                assert!(!generated_ids.contains(id))
-            }
-            for id in &generated_ids {
-                assert!(!self.generated_runtime_ids.contains(id));
-            }
-        }
-        self.generated_runtime_ids.extend(generated_ids);
+        self.invariants.save_generated_object_ids(generated_ids);
     }
 
     pub fn estimate_effects_size_upperbound(&self) -> usize {
@@ -707,15 +720,26 @@ impl<'backing> TemporaryStore<'backing> {
             .fold(0, |sum, obj| sum + obj.object_size_for_gas_metering())
     }
 
-    /// Validates gasless post-execution invariants, using the withdrawal reservations derived
-    /// from the input reservations cached by [`Self::set_invariant_inputs`].
+    /// Validates gasless post-execution requirements using the reservations cached when the store
+    /// is constructed.
     pub(crate) fn check_gasless_execution_requirements(&self) -> Result<(), String> {
-        self.check_gasless_execution_requirements_with_reservations(Some(
-            &self.invariants.gasless_reservations(),
-        ))
+        use sui_types::balance::Balance;
+
+        // Gasless requirements are expressed in coin types `T`, while the shared input reservation
+        // budget is keyed by accumulator types `Balance<T>`.
+        let withdrawal_reservations = self
+            .post_execution_check_inputs
+            .input_reservations
+            .iter()
+            .filter_map(|((owner, ty), amount)| {
+                Balance::maybe_get_balance_type_param(ty)
+                    .map(|coin_type| ((*owner, coin_type), *amount))
+            })
+            .collect();
+        self.check_gasless_execution_requirements_with_reservations(Some(&withdrawal_reservations))
     }
 
-    /// Validates gasless post-execution invariants:
+    /// Validates gasless post-execution requirements:
     /// - No new objects were created or existing objects mutated (written_objects is empty)
     /// - The set of deleted objects exactly equals the set of input Coin objects
     /// - Each recipient receives at least the minimum transfer amount per token type
@@ -723,7 +747,9 @@ impl<'backing> TemporaryStore<'backing> {
     ///
     /// Parameterized entry for the legacy path, which computes its (flag-gated) reservations
     /// out-of-band. Deleted with legacy at the next execution cut.
-    pub fn check_gasless_execution_requirements_with_reservations(
+    // TODO: This is kept as pub due to legacy callers. Once the legacy path is deleted in a
+    // newer executor, we can fold this into `check_gasless_execution_requirements`.
+    pub(crate) fn check_gasless_execution_requirements_with_reservations(
         &self,
         withdrawal_reservations: Option<&BTreeMap<(SuiAddress, TypeTag), u64>>,
     ) -> Result<(), String> {
@@ -875,19 +901,6 @@ impl<'backing> TemporaryStore<'backing> {
 
     pub fn protocol_config(&self) -> &'backing ProtocolConfig {
         self.protocol_config
-    }
-
-    /// Cache the transaction-derived inputs the system-invariant checks need. Must be called once,
-    /// before execution, after any gas-smash filtering of `gas_data`.
-    /// See [`invariants::InvariantChecker::set_transaction_inputs`].
-    pub(crate) fn set_invariant_inputs(
-        &mut self,
-        transaction_kind: &TransactionKind,
-        gas_data: &GasData,
-        transaction_signer: SuiAddress,
-    ) {
-        self.invariants
-            .set_transaction_inputs(transaction_kind, gas_data, transaction_signer);
     }
 
     /// Run the (read-only) SUI-conservation and balance-accumulator invariant checks.
@@ -1080,6 +1093,80 @@ impl RuntimeObjectResolver for TemporaryStore<'_> {
             epoch_id,
         )
     }
+}
+
+/// Compute the per-`(address, type)` funds-accumulator reservation budget authorized by the
+/// transaction. Today every funds accumulator is a `Balance<T>`, but the `(address, TypeTag)`
+/// keying lets this generalize as more accumulator types are added. Sources:
+/// - PTB `FundsWithdrawalArg`s for any supported accumulator type (sender or sponsor as owner).
+/// - Gas paid entirely from address balance (credits `(gas_owner, Balance<SUI>)`).
+/// - Gas-data entries with coin-reservation digests (also credit `(gas_owner, Balance<SUI>)`).
+fn compute_input_reservations(
+    transaction_kind: &TransactionKind,
+    gas_data: &GasData,
+    transaction_signer: SuiAddress,
+    enable_gasless: bool,
+) -> BTreeMap<(SuiAddress, TypeTag), u64> {
+    use sui_types::balance::Balance;
+    use sui_types::gas_coin::GAS;
+    use sui_types::transaction::{Reservation, WithdrawFrom, is_gas_paid_from_address_balance};
+
+    let is_gasless = enable_gasless && is_gasless_transaction(gas_data, transaction_kind);
+    let mut reservations: BTreeMap<(SuiAddress, TypeTag), u64> = BTreeMap::new();
+    let sui_balance_type = Balance::type_tag(GAS::type_tag());
+
+    for arg in transaction_kind.get_funds_withdrawals() {
+        let owner = match arg.withdraw_from {
+            WithdrawFrom::Sender => transaction_signer,
+            WithdrawFrom::Sponsor => gas_data.owner,
+        };
+        let Reservation::MaxAmountU64(reservation) = arg.reservation;
+        let entry = reservations
+            .entry((owner, arg.type_arg.to_type_tag()))
+            .or_insert(0);
+        *entry = entry.saturating_add(reservation);
+    }
+
+    // Gasless transactions charge no gas, so gas sources grant no reservation (their budget is
+    // validated to be 0 anyway; skipping keeps the map free of a phantom zero entry).
+    if !is_gasless && is_gas_paid_from_address_balance(gas_data, transaction_kind) {
+        let entry = reservations
+            .entry((gas_data.owner, sui_balance_type.clone()))
+            .or_insert(0);
+        *entry = entry.saturating_add(gas_data.budget);
+    }
+
+    for entry in &gas_data.payment {
+        if let Ok(parsed) = ParsedDigest::try_from(entry.2) {
+            let entry = reservations
+                .entry((gas_data.owner, sui_balance_type.clone()))
+                .or_insert(0);
+            *entry = entry.saturating_add(parsed.reservation_amount());
+        }
+    }
+
+    reservations
+}
+
+/// What each `Publish`/`Upgrade` command declares about the package it writes, in command order.
+/// `None` for transaction kinds that are not PTBs.
+fn declared_packages(
+    transaction_kind: &TransactionKind,
+) -> Option<Vec<(usize, BTreeSet<ObjectID>)>> {
+    let TransactionKind::ProgrammableTransaction(pt) = transaction_kind else {
+        return None;
+    };
+    Some(
+        pt.commands
+            .iter()
+            .filter_map(|command| match command {
+                Command::Publish(modules, dep_ids) | Command::Upgrade(modules, dep_ids, _, _) => {
+                    Some((modules.len(), dep_ids.iter().copied().collect()))
+                }
+                _ => None,
+            })
+            .collect(),
+    )
 }
 
 /// Compares the owner and payload of an object.
